responds to init message + change form of frame message
