+++ src/environment/controls.cpp
@@ -1,6 +1,7 @@
 //
 // Created by Emmet on 2/23/2023.
 //
+
 #include "controls.h"
 
 auto Controls::serialize() const -> json {
