+++ src/messageCreator.h
@@ -11,18 +11,17 @@
 
 #include "types.h"
 
+#include "environment/controls.h"
 #include "environment/environment.h"
 #include "environment/organism.h"
 
 using json = nlohmann::json;
 
-class StateSerializer {
-private:
-
+class MessageCreator {
 public:
-	static auto environmentBody(i32 tick, Environment & environment, std::vector<Organism> & organisms) -> json;
+	static auto frameMessage(json && environmentBody) -> json;
 
-	static auto environmentFrameMessage(json && body) -> json;
+	static auto initMessage(json && environmentBody, json && controlsBody) -> json;
 };
 
 #endif //ALIFE_SERIALIZER_H
